chore: rename file and folders TextInput, SelectInput and TextAreaInput
