+++ src/components/Inputs/SelectInput/index.ts
@@ -1,0 +1,1 @@
+export { SelectInput } from './SelectInput'
