+++ src/components/Inputs/TextAreaInput/index.ts
@@ -1,0 +1,1 @@
+export { TextAreaInput } from './TextAreaInput'
