+++ src/components/Inputs/TextInput/index.ts
@@ -1,0 +1,1 @@
+export { TextInput } from './TextInput'
