+++ src/components/Inputs/index.tsx
@@ -1,7 +1,7 @@
 export { HTTPMethodSelect } from './HTTPMethodSelect'
 export { NumberInput } from './NumberInput'
 export { MillisecondInput } from './MillisecondInput'
-export { SelectInput } from './SelectInput'
+export { SelectInput } from './SelectInput/SelectInput'
 export { TextInput } from './TextInput'
 export { TextAreaInput } from './TextAreaInput'
 export { Toggle } from './Toggle'
